Crate telamon_capi

C API wrappers for calling Telamon through FFI.

The goal of the C API is to provide thin wrappers over existing Rust functionality, and only provides some cosmetic improvements to try and provide a somewhat idiomatic C interface.

Modules

error

explorer

ir
C API wrappers to create Telamon Kernels.

linalg
Linera algebra kernels.

search_space
C API wrappers to work with a Telamon search space.

Macros

exit

exit_if_null

unwrap_or_exit
Helper macro that unwraps a result. Exits with $error and sets the global ERROR variable when an error is encountered.

Structs

Context
Description of the evaluation context. In particular, in contains the mapping between argument names and argument values.

Device
Description of the targeted device.

Enums

DeviceId
Supported device types for running kernels.

KernelParameters
Supported kernels.

Traits

Kernel
A kernel that can be compiled, benchmarked and used for correctness tests.

Functions

env_logger_try_init
Initializes the logger.

kernel_free⚠
Deallocates kernel parameters created through one of the kernel_*_new functions. The params pointer becomes invalid and must not be used again after calling kernel_free.

kernel_matmul_new⚠
Instanciate a new kernel for matrix-matrix multiplication. The caller is responsible for deallocating the returned pointer using kernel_free. The tile_m, tile_n and tile_k parameters are read from during the call, but no pointer to the corresponding data is kept afterwards.

kernel_optimize⚠
Optimize a kernel on a given device. config_data points to a JSON-encoded string of length config_len containing the configuration parameters for the explorer.