Module telamon_capi::explorer

Functions

telamon_config_free⚠
Frees an explorer configuration.

telamon_config_new
Allocate a new explorer configuration object with suitable defaults.

telamon_explore⚠
Run the exploration according to the configuration.

telamon_string_copy⚠
Copy a C string pointer into a Rust String object. Use this to set string-valued configuration options.