Trait telamon_capi::Kernel

pub trait Kernel<'a> {
    type Parameters: Clone;
    type ExpectedOutput;
    fn name() -> &'static str;

    fn build_signature<AM>(
        parameters: Self::Parameters, 
        builder: &mut Builder<AM>
    ) -> Self
    where
        AM: ArgMap + Context + 'a;

    fn build_body(
        &self, 
        signature: &'b Signature, 
        ctx: &'b (Context + 'b)
    ) -> Vec<Candidate<'b>>;

    fn get_expected_output(&self, &Context) -> Self::ExpectedOutput;

    fn check_result(
        &self, 
        expected: &Self::ExpectedOutput, 
        context: &Context
    ) -> Result<(), String>;

    fn test_correctness<AM>(
        params: Self::Parameters, 
        num_tests: usize, 
        context: &mut AM
    )
    where
        AM: ArgMap + Context + 'a,
    { ... }

    fn test_bound<AM>(
        params: Self::Parameters, 
        num_tests: usize, 
        context: &mut AM
    ) -> Vec<BoundSample>
    where
        AM: ArgMap + Context + 'a,
    { ... }

    fn benchmark<AM>(
        config: &Config, 
        params: Self::Parameters, 
        num_samples: usize, 
        context: &mut AM
    ) -> Vec<f64>
    where
        AM: ArgMap + Context + 'a,
    { ... }

    fn deadend_ratio<AM>(
        params: Self::Parameters, 
        num_samples: usize, 
        context: &mut AM
    ) -> f64
    where
        AM: ArgMap + Context + 'a,
    { ... }
}

A kernel that can be compiled, benchmarked and used for correctness tests.

Associated Types

type Parameters: Clone

The input parameters of the kernel.

type ExpectedOutput

The values to expect as output.

Required Methods

fn name() -> &'static str

The name of the function computed by the kernel.

fn build_signature<AM>(
parameters: Self::Parameters,
builder: &mut Builder<AM>
) -> Self where
AM: ArgMap + Context + 'a,

Builds the signature of the kernel in the builder and returns an object that stores enough information to later build the kernel body and check its result. The is_generic flag indicates if th sizes should be instantiated.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn build_body(
&self,
signature: &'b Signature,
ctx: &'b (Context + 'b)
) -> Vec<Candidate<'b>>

Builder the kernel body in the given builder. This builder should be based on the signature created by build_signature.

fn get_expected_output(&self, &Context) -> Self::ExpectedOutput

Computes the expected output.

fn check_result(
&self,
expected: &Self::ExpectedOutput,
context: &Context
) -> Result<(), String>

Ensures the generated code performs the correct operation.

Provided Methods

fn test_correctness<AM>(
params: Self::Parameters,
num_tests: usize,
context: &mut AM
) where
AM: ArgMap + Context + 'a,

Generates, executes and tests the output of candidates for the kernel.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn test_bound<AM>(
params: Self::Parameters,
num_tests: usize,
context: &mut AM
) -> Vec<BoundSample> where
AM: ArgMap + Context + 'a,

Tests the correctness of the bound of kernels and returns the list of tested leafs along with the actual evaluation time.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn benchmark<AM>(
config: &Config,
params: Self::Parameters,
num_samples: usize,
context: &mut AM
) -> Vec<f64> where
AM: ArgMap + Context + 'a,

Runs the search and benchmarks the resulting candidate.

fn deadend_ratio<AM>(
params: Self::Parameters,
num_samples: usize,
context: &mut AM
) -> f64 where
AM: ArgMap + Context + 'a,

Computes the probability of encountering a dead-end when descending in the search tree.